requirements.txt
Django==2.2.16
pytest==6.2.4
pytest-pythonpath==0.7.3
pytest-django==4.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
Pillow==8.3.1
PyJWT==2.1.0
requests==2.26.0

pip~=22.1.2
toml~=0.10.2
wheel~=0.37.1
cryptography~=37.0.2
py~=1.11.0
setuptools~=60.2.0
djoser~=2.1.0
django-filter~=22.1